backend/requirements.txt
python-dotenv>=1.0.1
Django==5.2.1
djangorestframework==3.16.1
